requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
groq>=0.5.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
